+++ 02-dtmoney/src/App.tsx
@@ -3,6 +3,7 @@ import { Header } from "./components/Header";
 import { GlobalStyle } from "./styles/global"; // Importando o estilo global
 import Modal from 'react-modal';
 import { useState } from "react";
+import { NewTransactionModal } from "./components/NewTransactionModal";
 
 Modal.setAppElement('#root'); // Configurando o modal para ser aberto dentro do elemento root
 
@@ -19,12 +20,7 @@ export function App() { //permite que o componente seja exportado com o mesmo no
 		<> {/* O React precisa de um elemento raiz, no caso, o elemento <> </>  que é uma tag fragment */}
 			<Header onOpenNewTransactionModal={handleOpenNewTransactionModal}/> {/*Passando a função para o componente Header chamar o modal */ }
 			<Dashboard/>
-			<Modal 
-				isOpen={isNewTransactionModalOpen}
-				onRequestClose={handleCloseNewTransactionModal}
-			>
-				<h2>Cadastrar transação</h2>
-			</Modal>
+			<NewTransactionModal isOpen={isNewTransactionModalOpen} onClose={handleCloseNewTransactionModal} />			
 			<GlobalStyle/> {/* Usando o estilo global */}
 		</>
 	);