+++ docs/changelog.md
@@ -34,3 +34,4 @@
 - Terminado a aula 02 - Consumindo API - Configurando o MirageJS
 - Terminado a aula 03 - Consumindo API - Configurando o Axios
 - Terminado a aula 01 - Modal & Forms - Configurando o modal de criação
+- Terminado a aula 01 - Modal & Forms - Componente NewTransactionModal